PRJ: BIG CHANGES, VERIFICATION OF ANALYSIS INCOMPLETE API/ENH: the event number is now unsigned int long long with 8 byte BUG: cluster hit sets now the TDC number ENH: infinite scan parameters are supported now, the c++ libraries only have to know a scan parameter index labeling the different scan parameter settings MAINT: exception text formatting MAINT/ENH: new Basis::LongIntToStr function to print the event number to the console if needed API: Histogram::calculateThresholdScanArrays need the max/min scan parameters now, since it is now unknown to the histogramming class

+++ trunk/host/analysis/RawDataConverter/Clusterizer.h
@@ -154,7 +154,7 @@ private:
 	float _actualClusterCharge;											//temporary value holding the total charge value of the actual cluster
 
 	//actual event variables
-	unsigned int _actualEventNumber;  //event number value (unsigned long long: 0 to 18,446,744,073,709,551,615)
+	uint64_t _actualEventNumber;  										//event number value (unsigned long long: 0 to 18,446,744,073,709,551,615)
 	unsigned int _actualEventStatus;
 
 	bool _abortCluster;													//set to true if one cluster TOT hit exeeds _maxClusterHitTot, cluster is not added to the result array
